tests/wpt/web-platform-tests/css/css-images/parsing/image-rendering-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Images Module Level 3: parsing image-rendering with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-images-3/#propdef-image-rendering">
<meta name="assert" content="image-rendering supports the full grammar 'auto | smooth | high-quality | crisp-edges | pixelated'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("image-rendering", "auto");
test_valid_value("image-rendering", "smooth");
test_valid_value("image-rendering", "high-quality");
test_valid_value("image-rendering", "crisp-edges");
test_valid_value("image-rendering", "pixelated");
</script>
</body>
</html>
